* added extra method to check if a window is on the canvas * upgraded vs 2008 libraries * fixed console not closing * fixed double in game menu's

+++ src/actionmap.cpp
@@ -72,7 +72,7 @@ bool ActionMap::process (int key, bool down)
 
    Script& script = ScriptManager::getInstance().getTemporaryScript();
    script.prepareCall (actions[action]);
-   script.addParam ((int)down);
+   script.addParam(down);
    script.run(1);
    return true;
 }
